Artifact push URLs must redirect to the corresponding pull URL
This change is part of issue #112 ("Artifact configuration is confusing
and fragile, need canonical push/pull urls.")

It updates the bst-artifact-receive program to send a pull URL to
clients who access it over SSH.

This requires extra configuration in the artifact cache server, so that
it knows the correct pull URL. Versions of BuildStream which contain
this commit will not be able to communicate over SSH with artifact
caches that have not been updated to the same version.

Currently this is just used to double-check user configuration, but we
can now simplify the user facing configuration for artifact caches
completely.