Add support for multiple remote caches
This extends the 'artifacts' configuration block such that a list of
`url` mappings can be given instead of a single entry. For example:

    artifacts:
      - url: http://example.com/artifacts1
      - url: ssh://ostree@example.com/artifacts2

The OSTreeCache class is updated to set up multiple remotes and query
remote refs from all of them.

There are no automated tests for this yet.

Empty URLs ('') now raise an exception. They cause breakages internally
if we allow them through, and they can only occur if the user or our
tests are misconfiguring things somehow.

We report failure to fetch from the cache by printing a message to
stderr for now. This is because BuildStream's actual logging
functionality can't be used during frontend init -- see issue #168.