.gitlab-ci.yml: Use specific version of buildstream-fedora Docker image
The image is rebuilt nightly which means it could change under our
feet and cause confusing breakages. The reason for these nightly
rebuilds is so that the `bst-here` script offers a reasonably fresh
version of BuildStream to users. But it also means that dependencies
could change in broken ways.

You can see the list of tags here, ordered by newness:
https://hub.docker.com/r/buildstream/buildstream-fedora/tags/

The number in one of these tag names is the "version" of the image
(which is really the number of commits in the buildstream-docker-images
repo), while the hex string identifies the commit of BuildStream that
the image contains.

When changing the Docker image (to add new dependencies, or whatever)
you will need to wait until the new tag is pushed, then modify
.gitlab-ci.yml so that it pulls the new tag.