.gitlab-ci.yml: Use source distribution tarballs in all tests
This commit adds an initial stage to the pipeline to build a distribution
tarball and generate some helper scripts for later CI to use to unpack it
and install it.

Then, it makes sure that all pytest and integration test runs work from
the dist tarball instead of directly from the git repo, also the docs are
built from the dist tarball.

This ensures that everything continues to work with a dist tarball at all times.