tests/testutils/runcli.py: Always run with --no-colors
Pytest defaults to capturing stdout, which has the side effect of making
BuildStream realise it's not outputting to a terminal and disabling its
ANSI colour code output.

If you pass `--capture=no` then BuildStream detects the terminal and the
colour codes are emitted, which breaks all of the tests which match on
specific output from `bst`. Forcefully disable colours to fix that.