buildstream/plugins/elements/cmake.yaml: Always create build folder
This is the common practice with cmake
Actually, some modules will fail to build if this is not follow. For
example for llvm you wil get this error when configuring:

"
CMake would overwrite the makefiles distributed with LLVM.

Please create a directory and run cmake from there, passing the path
to this source directory as the last argument.
"