_ostree.py: Avoid raising GLib.Error from configure_remote()
This call can fail (if we pass an invalid remote name, for example) so
we should wrap the GLib.Error with OSTreeException.

In the case of GLib.Error this is especially important as GLib.Error
can't be pickled and so such exceptions disappear completely when we try
to propagate them back from multiprocessing subprocesses. See
<https://bugzilla.gnome.org/show_bug.cgi?id=791265>.