tests/frontend: Reverse ordering of cache specifications
The cache parameters are now in highest-to-lowest precidence
order corresponding to left-to-right along the line where they are
specified.

Once we make it possible to specify a list of URLs for each of these
parameters, the precidence will continue to always read from highest
to lowest corresponding to left-to-right since each list of URLs
will also read left-to-right.