Add support for version 1 workspaces.yml format
This introduces support for the version 1 workspaces.yml format. This
format allows extensibility at depths of the workspaces.yml config.

An example of this format, for a build element, alpha.bst, with two
source workspaces open is:

```
version: 1
build-elements:
   alpha.bst:
     sources:
       0:
         path: /workspaces/bravo
       1:
         path: /workspaces/charlie
```

Additional attributes at any depth are permitted, although these will be
ignored for a workspaces.yml parser that does not look for these.

As parts of the workspaces.yml are pruned as they are no longer
required, for instance, in cases where a workspace is closed, no
guarantee is given for the persistence of attributes that reside within
any part that may be pruned.

While the focus is now on the use of the version 1 workspaces.yml
format, the older (un-versioned) format is still supported for the
purposes of conversion to the version 1 workspaces.yml format. This
ensures that this change does not break workflows for those that upgrade
BuildStream with workspaces open.