meson: Make sure importing from the source tree still works
It's annoying and confusing to find that `import buildstream` fails when
you happen to be in a certain directory. This was happening because we
were importing the version number from a module generated by the Meson
build instructions. We can use a placeholder when running from the source
tree though.

This is also needed so that we can generate the man pages at build time,
because this requires being able to import buildstream.

The only catch is that when buildstream is imported from the current
directory, the version number is simply returned as '(uninstalled)' now.