utils.py: Add UtilError
Report UtilError instead of OSError and similar python errors.

Also ensure we catch system errors and raise UtilError with
descriptive text instead; for the user experience; this is the
difference between:

  o A FAILURE message with a description as to
    what went wrong (exception handled with UtilError)

  o A BUG message with the unhandled system error printed
    with a stack trace (exception left unhandled)

Also, UtilsError and ProgramNotFoundError are now public exceptions
declared in utils.py, where they will appear in the documentation.