Hack up initial helper for profiling the SafeHardlinks FUSE layer
This should work, but it doesn't because for some reason the subprocess
fails to write out a profile. If the equivalent process is run directly
from the commandline then it works fine no matter how we kill it. But
when run from the parent Python subprocess, sending it a signal causes
it to exit seemingly immediately.

All of SIGINT, SIGTERM and SIGHUP cause the same erroneous behaviour.
Explicitly handling SIGINT inside the subprocess also has no effect.
Trying to catching KeyboardInterrupt also doesn't have any effect.
However the process does exit with the expected returncode showing
that it was killed with whatever signal we used.

This makes no sense really and is probably not worth digging into
further.