plugins/sources/patch.py: Removing unneeded check
The patch plugin was checking if the target directory exists, however
this is automatically guaranteed by the Source abstract class and
documented to be guaranteed as well.

Since this error cannot be caught by the plugin (it will be caught
in advance by the Source class), removing the check from patch.py.