Documentation: Fix warnings, and make all warnings errors.
Some of the warnings from sphinx-build are really just warnings,
but a lot of the things we want to avoid and really break documentation,
like broken internal references and some invalid rst directives should
really be errors.

Now we treat all warnings as errors, this should ensure that
any commits landing upstream never break the docs.