utils.py: Sort the results of list_relative_paths()
This fixes a serious issue in the 'local' source plugin: this source
type generates its unique key from the list of filenames it pulls in
from the local machine, but the list would be unsorted so cache keys
would randomly vary.

Rather than fix this in the specific plugin, we chose to fix the
public utility function that generates the list of filenames to reduce
the chance of this mistake being repeated elsewhere.