utils.py: Make list_relative_paths() report sorted paths.
This is needed because plugins make use of this function to
generate a cache key which must be stable.

In addition to the above, this patch also mitigates the performance
hit of sorting, and allows _process_list() to function to iterate
through the generator when additional sorting is not needed.

This patch is an enhanced version of Sam Thursfield's patch on MR !216