Get version number w/o pkg_resources
This removes most of the usages of pkg_resources from the codebase, helping
the goal of getting rid of that completely.

With this change, version number is generated during install and embedded into
`__version__` which is then imported by root level `__init__`. From there,
it is used by other parts of the codebase when needed.

Generated `__version__` file is ignored and not tracked by git to prevent
unnecessary 'changes' messages and accidental commits of that file.