buildelement.py: Remove secret undocumented command lists
Originally this was created with also `bootstrap-commands` and
`test-commands` but these were never documented or used.

For `bootstrap-commands`, these originated in baserock but are
basically a part of the `configure-commands` stage, prepending
commands to `configure-commands` is more suitable here since we
now have the prepend/append list directives.

For `test-commands`, these were never used and it's unclear at
this time if it's the correct place for it. It would be interesting
to implement `test-commands` as a separate operation which can run
in parallel with reverse dependency builds (no need to block the
build of a reverse dependency on failing tests, we can still fail
the build as a whole based on a failing test without blocking
builds).