Add a 'path' config option to junction elements
This makes it possible to depend on a project which is in a subdirectory
of a Git repository.

The error message given when the expected project.conf file is not found
has also been improved. Previously the error would look like this:

    Error loading pipeline: Could not find file at
    /home/sam/.cache/buildstream/build/freedesktop-sdk-junction-rvmn17s2/project.conf

This is giving the path of an internal temporary directory where the foreign
project's source is checked out. The new error gives more information:

    Error loading pipeline: Could not find the project.conf file for junction
    element at freedesktop-sdk-junction.bst [line 1 column 0]. Expecting a
    project at path '.' within git source at freedesktop-sdk-junction.bst [line
    4 column 2]