git.py: Add the ability to set whether to checkout submodules
This can be set as broad as "No submodule in this element should be
checked out" and as narrow as "only this submodule should be checked
out"

Note that this will change the cache-keys of every element that uses a
git source. I could have structured the code to prevent the cache key
changing, but that adds extra complexity in the long-term.