tests/integration/shell.py: Dont use shlex, use command vectors
Using shlex messes with the argv we want to pass to a shell, this
was not working for the newly added test which passes "${var}" strings
as arguments to the shell.

Also, removed a redundant line to explicitly build something
in the 'no_shell' test; that is taken care of by the helper.