Add the FixedText and WallclockTime widgets
WallclockTime is set at the time a message is created. Since messages
can be placed in queues before being rendered, this is not always
the same time as the wallclock time when rendering the message, so it
needs to be stored in the message.

widget.py: Add the two new widgets
_message.py: record system time when a message is created.