_versions.py: Added new file just to hold basic symbolic versions
This is where we're going to store BST_CORE_ARTIFACT_VERSION
and BST_FORMAT_VERSION from now on.

This was a little bit weirdly defined before, and now we need to
at least have BST_FORMAT_VERSION available from the frontend/cli
without onerous imports (we need to import it there, but dont want
to slow down load time for bash completion runs).