Serialize workspace modifications in the main process.
This patch makes the child process send any workspace state back
to the main process and saves it there directly before processing
the result of a build job, such that workspace modifications in
child tasks work transparently and no special care needs to be taken
to save them except when doing so in the main process.

Also this removes the line where we update workspace data at staging
time.

This fixes issue #352