_frontend/cli.py, _pipeline.py: Add options for cross junction tracking.
This patch makes cross junction tracking disabled by default, which
was the initial intention when landing project.refs but never got around
to doing this (intended to get addressing of junctioned elements via
command line sorted first, but didnt happen).

This adds the following options to enable cross-junction tracking:

  o bst build -J / --track-cross-junctions

  o bst fetch -J / --track-cross-junctions

  o bst track -J / --cross-junctions

This also fixes `bst fetch --track` which had a bug, it was avoiding
to track and fetch elements which are in a `cached` consistency state,
which is wrong when `--track` is specified.

This also updates some test cases which were broken by
this change.

This fixes issue #354