frontend: Log which elements resolved new keys
It can be difficult to pick out important information in all the messages,
so it's convenient to have classes of them collected into one place.

One such class of important information is when the cache key of an element changes,
such as when element's source is tracked or cached from a workspace,
since it can be important to know what the cache key is for that artifact.

Fixes issue #252