source.py, element.py, _pipeline.py: Streamling preflighting.
Instead of having the pipeline preflight all sources separately
from elements, have the element preflight it's sources.

This is in order to simplify the shared code path for the pipeline
and the loader to use for instantiating elements.

Also updated tests to expect the new ElementError and SourceError
instead of the PipelineError which was raised for preflighting before.