Overwrite sys.excepthook to handle exceptions in the main application.
This uses the existing messaging system to report errors in the main
application as a BUG type. It requires the use of a global_app variable
which isn't ideal; this may be replaced in future. This partially addresses
issue #197.

Theoretically, an exception could occur before Scheduler.loop is set up,
hence the check for it when terminating all jobs.

NOTE: This was originally submitted by Jim MacArthur, and manually
      reapplied after some refactoring took place.