_frontend/cli.py: Fetch all elements when tracking a build plan
It makes not sense to type `bst fetch --track --deps all <targets>`,
because tracking will inevitably modify the build plan.

Stream initialization will not cope with this either, instead of
silently doing something which does not make any sense, we add an
assertion that this should not happen.

Unfortunately since `plan` is the default deps type for `bst fetch`,
this is likely to happen so it's important to correct.

This patch adds a warning in the case tracking of the build plan
elements is requested, and converts the request to track all elements
instead.