Handle cross junction elements in workspaces.
Workspaces are now index by colon separated junction path. This
now allows to create workspaces for elements in external projects.

Workspaces are owned by context instead of root project. However
it is initialized once top-level project is registered as we need
to resolve paths relatively to this top-level project.

Part of #359.