_loader/loader.py: Report element-path when failing to load elements
It can be confusing, especially to new BuildStream users, that the CLI
expects targets to be specified relative to element-path and not the
current directory. Previously, the CLI would give a generic message
stating that the file could not be found but it was not obvious that it
was looking in the `element-path` directory.

Explicitly print the element-path in the summary. Also, try to check if
the specified element exists in the elements directory and print a hint
to use the element-path relative paths if that's the case.

Fixes #396.

This is is also related to #341. This commit aims to tackle that issue
by trying to educate users about element-path.