doc: Improve the new console output html generator
Before we were creating one description file for each output,
making it easier to declare a make rule for it - but the result
was that we would have to build things more and it takes a
long time.

Instead, now we have session files which describe a series of commands
to run in a session, and each command optionally produces an output file.