doc/bst2html.py: Add support for `fake-output` when running commands.
When specifying a fake-output string, we don't really run the command
or assume it was a `bst` command, and we pretend that `fake-output`
was the output of the command.

Specifying an empty string explicitly enables the behavior too
for faking a command that has no stdout/stderr.

This also adds the "remove-files" hack allowing the session scripts
to remove files before executing commands (kind of unsure if we're
gonna keep this...)