Restrict version of pylint
pylint >2 is not compatible with pytest_pylint in its current form. As
such, allowing any version of pylint for testing results in a failure.

This commit restricts down the allowable versions of pylint to those
that are both compatible with pytest_pylint, and also offer the feature
set that we require.

See https://gitlab.com/BuildStream/buildstream/issues/427 for further
details.